better things

+++ update.sh
@@ -5,31 +5,34 @@ brew outdated||\
     read -p 'Would you like to upgrade Brew (y/n): ' upgradeBrew
 if [[ "$upgradeBrew" == "y" ]]; then
     brew upgrade
+    echo "brew cask upgrade"
+    brew cask upgrade
 fi
 
-echo "brew cask upgrade"
-brew cask upgrade
 echo 
 echo "vim"
 
-vim +PluginUpdate +qall
+nvim +PlugUpdate +qall
 
-echo 
-echo "conda update conda"
-conda update conda
-echo "conda update anaconda"
-conda update anaconda
-echo 
-echo "conda update conda"
-conda update conda
-echo "conda update anaconda"
-conda update anaconda
-echo 
-echo "conda update --all"
-conda update --all
-
-echo "Upgrading pip3"
-pip3 list --outdated --format=freeze | grep -v '^\-e' | cut -d = -f 1  | xargs -n1 -I {} sh -c "echo \"Updating {}\" && pip3 install -U {}"
+read -p 'Would you like to upgrade Python: ' upgradePython
+if [[ "$upgradePython" == "y" ]]; then
+    echo 
+    echo "conda update conda"
+    conda update conda
+    echo "conda update anaconda"
+    conda update anaconda
+    echo 
+    echo "conda update conda"
+    conda update conda
+    echo "conda update anaconda"
+    conda update anaconda
+    echo 
+    echo "conda update --all"
+    conda update --all
+    
+    echo "Upgrading pip3"
+    pip3 list --outdated --format=freeze | grep -v '^\-e' | cut -d = -f 1  | xargs -n1 -I {} sh -c "echo \"Updating {}\" && pip3 install -U {}"
+fi
 
 echo "npm update -g --verbose"
 npm update -g --verbose
